moving to upppercase directories
