+++ CLI/Modules/GetPenaltyTermModule.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include "yaml-cpp/yaml.h"
-#include "cli/api/plugin.hpp"
+#include "CLI/API/plugin.hpp"
 
 namespace M3{
