+++ CMakeLists.txt
@@ -295,8 +295,7 @@ endforeach()
 add_library(MaCh3 INTERFACE)
 target_link_libraries(MaCh3 INTERFACE MaCh3CompilerOptions MaCh3GPUCompilerOptions)
 # Build components
-# add_subdirectory(api)
-add_subdirectory(cli)
+add_subdirectory(CLI)
 add_subdirectory(Manager)
 add_subdirectory(Parameters)
 add_subdirectory(Splines)
