+++ Diagnostics/CMakeLists.txt
@@ -1,36 +1,3 @@
-# set(HEADERS
-#     ProcessMCMCPlugin.hpp
-# )
-
-
-# add_library(CorePlugins STATIC
-#     ProcessMCMCPlugin.cpp
-#     DiagMCMCPlugin.cpp
-#     GetPenaltyTermPlugin.cpp
-# )
-
-# target_link_libraries(CorePlugins PRIVATE MaCh3::All)
-# target_include_directories(CorePlugins PUBLIC $<BUILD_INTERFACE:${CMAKE_BINARY_DIR}/include>)
-
-# set_target_properties(Plugins PROPERTIES
-#     PUBLIC_HEADER "${HEADERS}"
-#     EXPORT_NAME Plugins)
-
-
-# install(TARGETS Plugins
-#         EXPORT MaCh3-targets
-#         LIBRARY DESTINATION lib/
-#         PUBLIC_HEADER DESTINATION include/Plugins)
-
-# add_library(MaCh3::CorePlugins ALIAS CorePlugins)
-
-# set(CLI_MODULES
-#     ProcessMCMCPlugin.cpp
-#     DiagMCMCPlugin.cpp
-#     GetPenaltyTermPlugin.cpp
-# )
-
-# list(TRANSFORM CLI_MODULES PREPEND "../cli/modules/")
 
 add_custom_target(DiagApps)
